Draw bevel buttons to an offscreen bitmap at the full size of the widget, then draw the bitmap to our real surface with the desired size and position within the bitmap.
The problem is that HITheme doesn't allow us to draw *parts*
of a control, so if in one expose event we draw the top 25% of
a bevel button (while the rest is obscured by a scrolled window
viewport) then HITheme thinks that is the full height and in
the case of the bevel button it renders the full-height gradient
in only 25% of the height. Then when we expose the rest of the
button, HITheme draws the remaining 75% of the button with the
full-height gradient and the result is that we have these two
gradients instead of a single one.

https://bugzilla.xamarin.com/show_bug.cgi?id=25715